Several updates to the heat step:
* Update it to be HSV based internally to allow easier color shifts
* Link up color shifting to a control
* Link saturation shifting to a control
* Link up heating and cooling to control
* Added HSV to RGB conversions to the Base class
* Fixed devserver max color value

+++ floor/floor/controller/controller.py
@@ -61,7 +61,7 @@ class Controller(object):
         :param factor: a scaling factor from 0.0 to 1.0
         :return: none
         """
-        new_max = factor * self.max_led_value
+        new_max = int(factor * self.max_led_value)
         self.processor.set_max_value(new_max)
 
     def square_weight_on(self, index):
